requirements.txt
# 核心依赖
baostock>=0.8.8
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0

# 新增：Akshare数据源
akshare>=1.12.0

# 其他依赖
pyyaml>=6.0
loguru>=0.5.0  # 可选，更友好的日志